Update Docker Swarm-etcd environment
Update the etcd-backed Docker Swarm environment with new methods/techniques learned in building the Vagrant environment. This involves changes to the external YAML file (renamed to "machines.yml") and the Vagrantfile. Only minor changes were made to the ERB template and the provisioning script (to eliminate a dependency on shared folders).

Signed-off-by: Scott S. Lowe <scott.lowe@scottlowe.org>